Use Math functions to determine the line number string length
